desktop_app/requirements.txt
requests
websockets
Pillow
PyQt6
qrcode[pil]
reportlab
